+++ httpx/_client.py
@@ -10,7 +10,6 @@ from ._config import (
     DEFAULT_MAX_REDIRECTS,
     DEFAULT_TIMEOUT_CONFIG,
     UNSET,
-    EventHooks,
     Limits,
     Proxy,
     Timeout,
@@ -67,10 +66,12 @@ class BaseClient:
         cookies: CookieTypes = None,
         timeout: TimeoutTypes = DEFAULT_TIMEOUT_CONFIG,
         max_redirects: int = DEFAULT_MAX_REDIRECTS,
-        event_hooks: dict = None,
+        event_hooks: typing.Dict[str, typing.List[typing.Callable]] = None,
         base_url: URLTypes = "",
         trust_env: bool = True,
     ):
+        event_hooks = {} if event_hooks is None else event_hooks
+
         self._base_url = self._enforce_trailing_slash(URL(base_url))
 
         self._auth = self._build_auth(auth)
@@ -79,7 +80,10 @@ class BaseClient:
         self._cookies = Cookies(cookies)
         self._timeout = Timeout(timeout)
         self.max_redirects = max_redirects
-        self._event_hooks = EventHooks(event_hooks or {})
+        self._event_hooks = {
+            "request": list(event_hooks.get("request", [])),
+            "response": list(event_hooks.get("response", [])),
+        }
         self._trust_env = trust_env
         self._netrc = NetRCInfo()
 
@@ -93,7 +97,9 @@ class BaseClient:
         return url.copy_with(path=url.path + "/")
 
     def _get_proxy_map(
-        self, proxies: typing.Optional[ProxiesTypes], allow_env_proxies: bool,
+        self,
+        proxies: typing.Optional[ProxiesTypes],
+        allow_env_proxies: bool,
     ) -> typing.Dict[str, typing.Optional[Proxy]]:
         if proxies is None:
             if allow_env_proxies:
@@ -121,12 +127,17 @@ class BaseClient:
         self._timeout = Timeout(timeout)
 
     @property
-    def event_hooks(self) -> EventHooks:
+    def event_hooks(self) -> typing.Dict[str, typing.List[typing.Callable]]:
         return self._event_hooks
 
     @event_hooks.setter
-    def event_hooks(self, event_hooks: dict) -> None:
-        self._event_hooks = EventHooks(event_hooks)
+    def event_hooks(
+        self, event_hooks: typing.Dict[str, typing.List[typing.Callable]]
+    ) -> None:
+        self._event_hooks = {
+            "request": list(event_hooks.get("request", [])),
+            "response": list(event_hooks.get("response", [])),
+        }
 
     @property
     def auth(self) -> typing.Optional[Auth]:
@@ -520,7 +531,7 @@ class Client(BaseClient):
         limits: Limits = DEFAULT_LIMITS,
         pool_limits: Limits = None,
         max_redirects: int = DEFAULT_MAX_REDIRECTS,
-        event_hooks: dict = None,
+        event_hooks: typing.Dict[str, typing.List[typing.Callable]] = None,
         base_url: URLTypes = "",
         transport: httpcore.SyncHTTPTransport = None,
         app: typing.Callable = None,
@@ -684,7 +695,10 @@ class Client(BaseClient):
             cookies=cookies,
         )
         return self.send(
-            request, auth=auth, allow_redirects=allow_redirects, timeout=timeout,
+            request,
+            auth=auth,
+            allow_redirects=allow_redirects,
+            timeout=timeout,
         )
 
     def send(
@@ -714,7 +728,10 @@ class Client(BaseClient):
         auth = self._build_request_auth(request, auth)
 
         response = self._send_handling_redirects(
-            request, auth=auth, timeout=timeout, allow_redirects=allow_redirects,
+            request,
+            auth=auth,
+            timeout=timeout,
+            allow_redirects=allow_redirects,
         )
 
         if not stream:
@@ -1133,7 +1150,7 @@ class AsyncClient(BaseClient):
         limits: Limits = DEFAULT_LIMITS,
         pool_limits: Limits = None,
         max_redirects: int = DEFAULT_MAX_REDIRECTS,
-        event_hooks: dict = None,
+        event_hooks: typing.Dict[str, typing.List[typing.Callable]] = None,
         base_url: URLTypes = "",
         transport: httpcore.AsyncHTTPTransport = None,
         app: typing.Callable = None,
@@ -1298,7 +1315,10 @@ class AsyncClient(BaseClient):
             cookies=cookies,
         )
         response = await self.send(
-            request, auth=auth, allow_redirects=allow_redirects, timeout=timeout,
+            request,
+            auth=auth,
+            allow_redirects=allow_redirects,
+            timeout=timeout,
         )
         return response
 
@@ -1329,7 +1349,10 @@ class AsyncClient(BaseClient):
         auth = self._build_request_auth(request, auth)
 
         response = await self._send_handling_redirects(
-            request, auth=auth, timeout=timeout, allow_redirects=allow_redirects,
+            request,
+            auth=auth,
+            timeout=timeout,
+            allow_redirects=allow_redirects,
         )
 
         if not stream:
@@ -1422,7 +1445,9 @@ class AsyncClient(BaseClient):
                 history.append(response)
 
     async def _send_single_request(
-        self, request: Request, timeout: Timeout,
+        self,
+        request: Request,
+        timeout: Timeout,
     ) -> Response:
         """
         Sends a single request, without handling any redirections.
